packages/project-chat-gb/project_chat_gb-0.0.1-py3-none-any.whl/server_part/common/utils.py
"""functions used both by clients and server"""
import json
from socket import socket
from src.server_part.common.variables import MAX_PACKAGE_LENGTH, ENCODING
import logging

logger = logging.getLogger('app.server')


# @log
def get_message(sock) -> dict:
    """
    The utility for receiving and decoding messages
    recives bytes, returns dict
    """
    encoded_response = sock.recv(MAX_PACKAGE_LENGTH)
    logger.debug(encoded_response)
    if isinstance(encoded_response, bytes):
        json_response = encoded_response.decode(ENCODING)
        response = json.loads(json_response)
        if isinstance(response, dict):
            return response
        raise ValueError
    raise ValueError


# @log
def send_message(sock: socket, message: dict):
    """
    The utility encoding messages to json
    end sending to the passed socket
    """
    js_message = json.dumps(message)
    encoded_message = js_message.encode(ENCODING)
    sock.send(encoded_message)
